requirements.txt
absl-py==2.3.1
accelerate==1.6.0
ace_step @ git+https://github.com/ace-step/ACE-Step.git@6ae0852b1388de6dc0cca26b31a86d711f723cb3
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
blis==1.3.0
Brotli==1.1.0
catalogue==2.0.10
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
comm==0.2.3
confection==0.1.5
contourpy==1.3.2
cutlet==0.5.0
cycler==0.12.1
cymem==2.0.11
datasets==3.4.1
debugpy==1.8.15
decorator==5.2.1
diffusers==0.34.0
dill==0.3.8
docopt==0.6.2
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.12.0
fugashi==1.5.1
gradio==5.38.0
gradio_client==1.11.0
groovy==0.1.2
grpcio==1.73.1
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.0
ipython==8.37.0
jaconv==0.4.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
lightning-utilities==0.14.3
llvmlite==0.44.0
loguru==0.7.3
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mojimoji==0.0.13
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
nest-asyncio==1.6.0
networkx==3.4.2
num2words==0.5.14
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
orjson==3.11.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
peft==0.16.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py3langid==0.3.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.3
pypinyin==0.53.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruff==0.12.4
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.2
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.7.4
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
triton==3.2.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unidic-lite==1.0.8
urllib3==2.5.0
uvicorn==0.35.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
hydra-core
